port Kollect instances to its own file
